chore: Moves hook from layout and attempts to use isValidating
