+++ packages/core/src/components/region/RegionModal/useRegionModal.ts
@@ -1,0 +1,33 @@
+import { useUI } from '@faststore/ui'
+import { deliveryPromise } from 'discovery.config'
+import { useEffect, useRef } from 'react'
+import { sessionStore, useSession } from 'src/sdk/session'
+
+export function useRegionModal() {
+  const { openModal } = useUI()
+  const { isValidating } = useSession()
+
+  // Ref to track the previous value of isValidating
+  const prevIsValidating = useRef(isValidating)
+
+  const openRegionModal = () => {
+    const { postalCode } = sessionStore.read()
+    if (!postalCode) {
+      openModal()
+    }
+  }
+
+  // Effect to handle when isValidating changes from true to false
+  useEffect(() => {
+    if (prevIsValidating.current && !isValidating) {
+      if (deliveryPromise.enabled && deliveryPromise.mandatory) {
+        openRegionModal()
+      }
+    }
+
+    // Update the previous value of isValidating
+    prevIsValidating.current = isValidating
+  }, [openRegionModal])
+
+  return openRegionModal
+}
